refactor: Remirror link extension
